2.2.0 for Minecraft 1.21.11 feat: minor `/sba` search improvement feat: log4j config for just show SBA's debug messages on dev env

+++ src/main/java/com/fix3dll/skyblockaddons/core/render/chroma/ManualChromaManager.java
@@ -49,7 +49,7 @@ public class ManualChromaManager {
             x *= featureScale;
             y *= featureScale;
         }
-        int scale = (int) MC.getWindow().getGuiScale();
+        int scale = MC.getWindow().getGuiScale();
         x *= scale;
         y *= scale;
